Refactor: Organize app into core, features, and shared modules
