+++ lib/core/constants/app_colors.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+class AppColors {
+  static const Color kContentColor = Color(0xfffdfdfd);
+  static const Color kscaffoldBackgroundColor = Color(0xfff1f0f5);
+  static const Color kprimaryColor = Color(0xffcb0a09);
+  static const Color kTextAndIconColor = Color(0xFF1C3941);
+  static const Color kunselectedItemColor = Color(0xFF8E8D90);
+}
