+++ lib/core/providers/locale_provider.dart
@@ -1,0 +1,36 @@
+import 'package:flutter/material.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+
+class LocaleProvider with ChangeNotifier {
+  Locale _locale;
+  final SharedPreferences prefs;
+
+  LocaleProvider(this.prefs)
+      : _locale = Locale(prefs.getString('language_code') ?? 'ar') {
+    _loadLocale();
+  }
+
+  Locale get locale => _locale;
+
+  Future<void> _loadLocale() async {
+    final languageCode = prefs.getString('language_code') ?? 'ar';
+    _locale = Locale(languageCode);
+    notifyListeners();
+  }
+
+  void toggleLocale() {
+    if (_locale.languageCode == 'en') {
+      _locale = const Locale('ar', 'AE');
+    } else {
+      _locale = const Locale('en', 'US');
+    }
+    prefs.setString('language_code', _locale.languageCode);
+    notifyListeners();
+  }
+
+  void setLocale(Locale locale) {
+    _locale = locale;
+    prefs.setString('language_code', locale.languageCode);
+    notifyListeners();
+  }
+}
