+++ lib/features/home/cart_page.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class CartPage extends StatelessWidget {
+  const CartPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(
+        child: Text('Cart Page'),
+      ),
+    );
+  }
+}
