+++ lib/features/home/clients_page.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class ClientsPage extends StatefulWidget {
+  const ClientsPage({super.key});
+
+  @override
+  State<ClientsPage> createState() => _ClientsPageState();
+}
+
+class _ClientsPageState extends State<ClientsPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(),
+    );
+  }
+}
