+++ lib/features/home/favorites_page.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class FavoritesPage extends StatelessWidget {
+  const FavoritesPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(
+        child: Text('Favorites Page'),
+      ),
+    );
+  }
+}
