+++ test/widget_test.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+import 'package:jeb_ui/main.dart';
+
+void main() {
+  TestWidgetsFlutterBinding.ensureInitialized();
+
+  testWidgets('App boots to bottom navigation', (tester) async {
+    SharedPreferences.setMockInitialValues({});
+    final prefs = await SharedPreferences.getInstance();
+
+    await tester.pumpWidget(MyApp(prefs: prefs));
+    await tester.pumpAndSettle();
+
+    expect(find.byType(Scaffold), findsWidgets);
+    expect(find.byIcon(Icons.home), findsOneWidget);
+  });
+}
